requirements.txt
# ByteWallet AI — dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.3
httpx==0.28.1
python-dotenv==1.0.1
joblib==1.4.2
kaggle==1.7.4.2
pytest==8.3.5
pytest-asyncio==0.25.3
# Phase 1 — NL Function Calling
google-generativeai>=0.8.0
# Phase 2 — RAG on Transactions
chromadb>=0.6.0
sentence-transformers>=3.4.0
